basic cleanup

+++ models.py
@@ -33,5 +33,5 @@ t_data = Table(
     Column('id', BigInteger, nullable=False, unique=True, server_default=text("nextval('data_id_seq'::regclass)")),
     Column('beacon_id', ForeignKey('beacon.id'), nullable=False, index=True, server_default=text("nextval('data_beacon_id_seq'::regclass)")),
     Column('timestamp', Time),
-    Column('counter', Integer, server_default=text("0"))
+    Column('counter', BigInteger, server_default=text("0"))
 )
